GROOVY-9455: !(x instanceof T) shouldn't propagate T as inferred type
(cherry picked from commit d5e217aa6fa17b5b015a9f7120ba77817ddbb127)
